Separate ResultCollector base
Move the base ResultCollector class into its own file, rather than
embedding it in the source files for the BulletResultCollector subclass.
This will make it possible to add other implementations. Also, remove
unneeded overload of toVector3d.
