+++ drake/systems/plants/collision/BulletResultCollector.cpp
@@ -8,22 +8,12 @@ using namespace std;
 using namespace Eigen;
 
 namespace DrakeCollision {
-Vector3d toVector3d(const Vector3d& vec) { return vec; }
-
 Vector3d toVector3d(const btVector3& bt_vec) {
   Vector3d vec;
   vec(0) = (double)bt_vec.getX();
   vec(1) = (double)bt_vec.getY();
   vec(2) = (double)bt_vec.getZ();
   return vec;
-}
-
-void ResultCollector::addPointPairResult(const PointPair& result) {
-  pts.push_back(result);
-}
-
-PointPair ResultCollector::minDistPoint() {
-  return *min_element(pts.begin(), pts.end());
 }
 
 btScalar BulletResultCollector::addSingleResult(
