+++ drake/systems/plants/collision/BulletResultCollector.h
@@ -2,30 +2,10 @@
 #define DRAKE_SYSTEMS_PLANTS_COLLISION_BULLETRESULTCOLLECTOR_H_
 
 #include "drake/systems/plants/collision/DrakeCollision.h"
+#include "drake/systems/plants/collision/ResultCollector.h"
 
 namespace DrakeCollision {
-Eigen::Vector3d toVector3d(const Eigen::Vector3d& vec);
 Eigen::Vector3d toVector3d(const btVector3& bt_vec);
-
-class ResultCollector {
- public:
-  virtual ~ResultCollector() {}
-
-  virtual void addPointPairResult(const PointPair& result);
-
-  inline void addSingleResult(const ElementId idA, const ElementId idB,
-                              const Eigen::Vector3d& ptA,
-                              const Eigen::Vector3d& ptB,
-                              const Eigen::Vector3d& normal, double distance) {
-    addPointPairResult(PointPair(idA, idB, ptA, ptB, normal, distance));
-  }
-
-  std::vector<PointPair> getResults() const { return pts; }
-
-  PointPair minDistPoint();
-
-  std::vector<PointPair> pts;
-};
 
 class BulletResultCollector : public ResultCollector,
                               public btCollisionWorld::ContactResultCallback {
@@ -50,8 +30,6 @@ class BulletResultCollector : public ResultCollector,
   int curr_bodyA_idx;
   int curr_bodyB_idx;
 };
-
-typedef std::shared_ptr<ResultCollector> ResultCollShPtr;
 }
 
 #endif  // DRAKE_SYSTEMS_PLANTS_COLLISION_BULLETRESULTCOLLECTOR_H_
